requirements.txt
aldryn-client==2.3.5
appdirs==1.4.4
arrow==0.15.8
asgiref==3.2.10
astroid==2.4.2
attrs==18.2.0
autopep8==1.5.3
beautifulsoup4==4.8.2
binaryornot==0.4.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-aliases==1.0.1
cookiecutter==1.7.2
distlib==0.3.1
divio-cli==3.8.0
Django==3.0.8
django-countries==6.1.2
django-extensions==3.0.3
django-ipware==3.0.0
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
filelock==3.0.12
flake8==3.8.3
html5lib==1.1
idna==2.7
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
l18n==2018.5
lazy-object-proxy==1.4.3
longclaw==1.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
pbr==5.4.5
Pillow==7.2.0
pluggy==0.13.1
poyo==0.5.0
pyaml==18.11.0
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylint==2.5.3
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==3.13
requests==2.24.0
rope==0.17.0
six==1.12.0
snowballstemmer==2.0.0
soupsieve==2.0.1
sqlparse==0.3.1
stevedore==3.2.0
tabulate==0.8.3
text-unidecode==1.3
toml==0.10.1
ujson==1.35
Unidecode==1.1.1
urllib3==1.24.1
virtualenv==20.0.27
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wagtail==2.9.3
webencodings==0.5.1
Willow==1.3
wrapt==1.12.1
XlsxWriter==1.2.9
yapf==0.30.0
